refactor: remove hook suffix from hook files naming
