+++ apps/web/src/components/auth/SignUpForm.tsx
@@ -13,7 +13,7 @@ import { PasswordInput, Button, colors, Input, Text, Checkbox } from '../../desi
 import { GitHub } from '../../design-system/icons';
 import { API_ROOT, IS_DOCKER_HOSTED } from '../../config';
 import { applyToken } from '../../store/use-auth-controller';
-import { useAcceptInvite } from './use-accept-invite.hook';
+import { useAcceptInvite } from './use-accept-invite';
 import { useVercelParams } from '../../hooks/use-vercel-params';
 import { PasswordRequirementPopover } from './PasswordRequirementPopover';
