+++ apps/web/src/components/auth/use-accept-invite.ts
