+++ apps/web/src/components/templates/SaveChangesModal.tsx
@@ -1,6 +1,6 @@
 import { Group } from '@mantine/core';
 import { Button, Title, Text, Modal } from '../../design-system';
-import { IForm } from './use-template-controller.hook';
+import { IForm } from './use-template-controller';
 import { errorMessage } from '../../utils/notifications';
 import { useFormContext } from 'react-hook-form';
