+++ apps/web/src/components/templates/TemplatePageHeader.tsx
@@ -4,8 +4,8 @@ import { ArrowLeft } from '../../design-system/icons';
 import { ActivePageEnum } from '../../pages/templates/editor/TemplateEditorPage';
 import { useEnvController } from '../../store/use-env-controller';
 import { When } from '../utils/When';
-import { useStatusChangeControllerHook } from './use-status-change-controller.hook';
-import { useTemplateController } from './use-template-controller.hook';
+import { useStatusChangeControllerHook } from './use-status-change-controller';
+import { useTemplateController } from './use-template-controller';
 
 const Header = ({ activePage, editMode }: { editMode: boolean; activePage: ActivePageEnum }) => {
   if (activePage === ActivePageEnum.SETTINGS) {
