+++ apps/web/src/components/templates/in-app-editor/EnableAvatarSwitch.tsx
@@ -1,7 +1,7 @@
 import { Group } from '@mantine/core';
 import { Control, useController } from 'react-hook-form';
 import { Text, Switch } from '../../../design-system';
-import { IForm } from '../use-template-controller.hook';
+import { IForm } from '../use-template-controller';
 
 export function EnableAvatarSwitch({
   name,
